Consider battery conservation mode on Lenovo IdeaPad laptops when the battery state is unknown.
Lenovo IdeaPad laptop models offer a battery conservation mode to limit battery charging to 55-60% of its capacity to improve battery life. It is particularly useful when the laptop runs on external power most of the time.

The Linux kernel module ideapad_laptop exposes this specific functionality over sysfs.

This commit makes uPower consider the battery conservation mode status (when available) improving the correctness of the battery status displayed. For instance, vanilla uPower would cause GNOME Desktop to constantly show status "Estimating..." when the battery is not charging or stops charging because of the conservation mode enabled. This patch will make this case displayed as "Not Charging" instead.