dbus: linux: support enabling battery charge limiting
This introduces all the bits required for a GUI to enable or disable
configured charge limits from hwdb for a battery. UPower now exposes if
a battery supports this feature and if it is enabled, enabling it is
done via the EnableChargeLimit DBus method in a battery device.

When enabling the charge limit feature UPower writes a
charge-threshold-enabled file in the UPOWER_HISTORY_DIR for the case
that the firmware does not remember charge threshold settings after a
reboot.